cuczhongchou/requirements.txt
Django==1.8.2
django-admin-bootstrapped==2.5.4
django-filter==0.10.0
djangorestframework==3.1.3
Markdown==2.6.2
MySQL-python==1.2.5
pycrypto==2.6.1
Pygments==2.0.2
requests==2.7.0
wechat==0.4.13
wheel==0.24.0
